requirements/dev.txt
-r prod.txt
preprocessing @ git+ssh://git@github.com/OriGenAI/preprocessing.git
proteus-runtime @ git+ssh://git@github.com/OriGenAI/proteus-runtime.git
black==22.1.0
pre-commit==2.9.3
pytest-mock==3.5.1
requests-mock==1.8.0
pytest-bdd==5.0.0